Add Gradle versions plugin and update some "safe" dependencies (#649)
With the versions plugin, we can run `./gradlew dependencyUpdates` to see which of our dependencies are not the latest version.  Many of these dependencies we may not want to update; e.g., they may be test dependencies against specific artifact versions containing certain annotations.  But it's still good to be able to generate a report.

I also updated a couple of dependencies that are "safe;" they only impact testing code, not anything that ships with NullAway or that NullAway depends on after deployment.