[JSpecify] Support @NullUnmarked. (#651)
This PR adds support for JSpecify 0.3.0's `@NullUnmarked` annotation in the following scopes:

- Packages
- Classes
- Methods

We do not yet support the module scope, since we still retain JDK 8 compatibility in NullAway.

Most of this PR is comprised of test cases for `@NullUnmarked` and its interactions with `@NullMarked` and out existing annotated/unannotated code flags. Unfortunately, the matrix of possible interactions is far too large for the test suite to be exhaustive.

Note that there are a few follow up cases to fix regarding the interaction of `@NullUnmarked`/`@NullMarked` and `AcknowledgeRestrictiveAnnotations`, but that requires a slightly more general change in `RestrictiveAnnotationHandler` (see #573).