Allow AcknowledgeRestrictiveAnnotations to work on fields (#652)
This PR causes `@Nullable` fields in "unannotated" code to still be considered nullable iff  `AcknowledgeRestrictiveAnnotations=true`, including when used as part of a mixed-nullness compilation unit (e.g. `@NullMarked` and `@NullUnmarked` code within the same top-level class). 

This partially addresses #573 (We still need special handling of JSpecify's annotations even when `AcknowledgeRestrictiveAnnotations` isn't set to true).

To implement this, we had to add an additional `Handler` callback to intercept checking of fields nullability during dataflow. This is analogous to our existing callback for method nullability during dataflow, but might be worth a quick sanity check for potential performance regressions.