Add serialization format version to fix serialization output (#688)
Adds a version for all outputs of serialization features. Format of serialized outputs may change over time. This version serialization helps users to keep track of changes or maintain compatibility with previous versions of outputs.

If `SerializeFixMetadata` is activated, a file with name `serialization_version.txt` will be created under the root directory for all serialization outputs containing the version as string.