Fix serialization field region computation bug fix (#689)
This PR fixes a bug in field region computation. The bug is in the computation of region for fields initialized through chain of  member selects. Please see the example below:

```java
class C{
  @Nullable Object val;
}
class B{
   C c = new C();
}
class A {
    B b = new B();
    Object f = b.c.val;
}
```

For expression `Object f = b.c.val;` which causes an `ASSIGN_NULLABLE` error (`val` is `@Nullable`) the computed region is `null`. This PR resolves this bug by correctly computing the region and setting `f` as the region where the error is reported.