Add checks for proper JSpecify generic type instantiations (#680)
This PR adds initial checks for JSpecify generic type validity.  In particular it adds checks that for a type instantiation `Foo<@Nullable Bar>`, `Foo` is actually declared to allow a `@Nullable` type variable in that position (e.g., it is declared as `class Foo<T extends @Nullable Object>`.

These checks are only performed if a new `-XepOpt:NullAway:JSpecifyMode` configuration is enabled, and it is off by default.  We want to merge these checks to the main branch to enable easier testing by users, but we do not want them to be on by default yet, as they require further vetting on real, large repos.