Preserve nullness checks in final fields when propagating nullness into inner contexts. (#703)
This resolves #699 by making sure that dataflow information about `@Nullable final` fields is
propagated down to the nullness store of lambdas, local classes, and anonymous inner classes,
at the time of declaration. We extend this support to chains of `@Nullable final` fields.

The key insight is that, given that the field is `final`, once we know it is non-null due to dataflow,
that fact will forever remain true, regardless of which other methods access the field. So, if at the
time the closure for a lambda or class is being created, we know that the field is non-null, then
we know it will remain non-null in all of its usages through that closure.