Add an initial annotations artifact (with sample @Initializer implementation) (#709)
The main purpose of this change is to introduce our own annotations artifact, which
can be leveraged to implement more general contracts and other annotations of
interest to NullAway.

Additionally, we start by having our own `@Initializer` annotation "implementation"
in this annotations jar. I believe Facebook/Meta's Eradicate is being sunset in favor of
Nullsafe (internally at Meta) and NullAway (for OSS), and either way recommending 
their annotation jar for NullAway is more of a historical artifact of nullness checking at 
Uber and our previous use of Infer/Eradicate than anything else. NullAway will still
acknowledge any annotation with simple name `@Initializer`, but now we can
recommend a canonical alternative which we can make sure remains supported.

Additionally, we take the opportunity to make our `@Initializer` valid only on method
declarations (which is the only place NullAway checks for it).

We use `annotations` rather than `annotation`, because that's the choice JSpecify
0.3.0 went with in the end, and I refuse to spend another year discussing that 😉