Revert #762 (Grade upgrade) and add CI checking of publishing flows (#767)
A recent upgrade to our gradle version (#762) silently broke our
publishing flows.

In particular, `publish` and `publishToMavenLocal` are broken on Java 8,
resulting in the following error:

```
Execution failed for task ':annotations:publishMavenPublicationToMavenLocal'.
> Failed to publish publication 'maven' to repository 'mavenLocal'
   > Invalid publication 'maven': artifact file does not exist: '/home/runner/work/NullAway/NullAway/annotations/build/publications/maven/module.json.asc'
```

See this [CI job
run](https://github.com/uber/NullAway/actions/runs/5261437721/jobs/9509542067?pr=767).
There are further errors when the release is signed.

Below, we revert the gradle upgrade, and also add a CI job running
`publishToMavenLocal` to prevent
future regressions (We can't trivially test signing these publications
in a secure manner, unfortunately,
so we skip that for now. If that specifically breaks in the future, we
could provide a test gpg key for signing
in CI).