Add tasks to run JDK 8 tests on JDK 11+ (#778)
This PR adds tasks to additionally run certain test suites on JDK 8,
even when we are building on a JDK 11+ JVM. The approach is described
[here](https://jakewharton.com/build-on-latest-java-test-through-lowest-java/).
This moves us towards only supporting building NullAway on JDK 11+,
while still targeting Java 8 (so NullAway can still be used on JDK 8).
This will help us with keeping build tooling and dependencies up to
date, as more and more tools and plugins do not support running on JDK
8.

Specifically, we run the core NullAway tests and the tests from
`guava-recent-unit-tests` on JDK 8. We do _not_ run JarInfer tests, as
we intend to remove the ability to run JarInfer on JDK 8 (see #777).
Also, we only create the relevant test tasks when the Error Prone
version being used is 2.4.0 (compatible with Java 8). We have one CI
job, Java 11 + EP 2.4.0, that exercises this scenario. I confirmed
locally that when running the `testJdk8` tasks, the tests are running on
a JDK 8 JVM.

Additionally, the PR contains a minor tweak that allows
`publishToMavenLocal` to work when Gradle is run on JDK 11+.