Update to Gradle 8.2.1 (#781)
We had to back out a previous Gradle upgrade, but I believe now things
have been fixed up enough that we can upgrade again. One issue with the
upgrade is that publishing a build will fail if signing is not
functional and the version is not a `SNAPSHOT`. In particular, with our
default config, `LOCAL` versions must be signed. (This was the
documented behavior before this Gradle upgrade but for some reason the
build wasn't failing.) So, publishing `LOCAL` builds without signing
requires disabling a setting via a Gradle property or env var. We update
the CI config and the docs to reflect this.

I tested publishing a release to Sonatype (without pushing to Maven
Central) from my laptop and it succeeded. I have not yet validated the
actual artifacts. Will hold off on landing this until we're sure it's
thoroughly tested.

Fixes #768