JSpecify: avoid crashes when encountering raw types (#792)
We check for the presence of a raw type and (for now) bail out of any
checking. Adding further support for JSpecify checking of raw types is a
low priority (not even sure anything needs to be done).

This fixes the initial crash from #791, though we may want to leave that
issue open for now until we split off issues for the other crashes as
needed.