Fix JDK compatibility issue in LombokHandler and introduce AstHelpersBackports (#795)
Fixes #794. Since the JDK interface change was made for JDK 17, I think
even if NullAway was built on and targeted JDK 11, we would still have
this problem. I added some hacky unit test coverage for this code so
we'd have a way to test for regressions in the future. (The new unit
test fails the `:nullaway:testJdk8` task without this change.)

We missed this problem earlier since we suppressed a warning from
`AstHelpersSuggestions`. Even though we cannot accept some of the
suggestions from that check (since we may not require a recent enough
version of Error Prone), the issues it flags can be serious. So, this PR
removes all other suppressions of that checker and fixes the warnings.
We introduce an `AstHelpersBackports` class to contain any backported
logic from `AstHelpers` to enable the fixes.