Make GenericsChecks methods static (#805)
Previously, some methods in `GenericsChecks` were static methods and
some were instance methods. We did not get much of any benefit from the
instance methods, as it would be painful to create a single
`GenericsChecks` object and thread it everywhere it is needed. So, this
PR just converts all `GenericsChecks` methods to be static and passes
parameters as needed (which isn't too painful).