Bump gradle maven publish plugin from 0.21.0 to 0.25.3 (#810)
[Changelog for
plugin](https://github.com/vanniktech/gradle-maven-publish-plugin/blob/cf02bca4185083bfc5d258e81e82939b5b2451a8/CHANGELOG.md)

We have to do some renaming as the `javaSourcesJar` configuration is now
just `sourcesJar`. Also, the `:nullaway:buildWithNullAway` previously
used a hacky way to refer to the current NullAway jar that broke with
this upgrade. Here we clean things up to refer to the current jar more
directly.

The plugin has better integration with Sonatype and will close the
staging repository after all artifacts are uploaded. There is an option
to automatically release as well, but I haven't enabled that for now (it
makes testing of release uploads difficult), so you still need to go to
https://oss.sonatype.org to finalize the release. Relatedly, apparently
we don't need `--no-daemon` and `--no-parallel` flags anymore when
cutting a release; updated the release docs accordingly.

I tested locally and confirmed that the exact same number of files get
included in the jar-infer-cli shadow jar, and that cutting a real
release successfully uploads to Maven Central.