Run tests on JDK 21 (#826)
It looks like most of our tests pass on JDK 21. Exceptions are the
JSpecify generics tests, which rely on an API that changed in JDK 21
(see #827), and the JarInfer tests, as it looks like WALA does not yet
support running on JDK 21 (see #829). Core NullAway support (ignoring
the experimental/WIP JSpecify mode) should be working. This PR adds test
configs so that we run tests on JDK 21 wherever possible. We update our
Jacoco and Mockito versions to make this work.

We also take the opportunity to move some common test configuration code
to the `nullaway.java-test-conventions.gradle` file, which previously
just held configuration for Jacoco.