Ignore code coverage for method executed non-deterministically in tests (#838)
Fixes #828 (hopefully)

There is some non-determinism in whether
`com.uber.nullaway.dataflow.AccessPath.IteratorContentsKey#equals` runs
or not during tests (see the code comment). The non-determinism will be
hard for us to fix, so instead, add an annotation that tells JaCoCo to
ignore coverage of this method (which is low risk as it's an `equals()`
method that I don't expect will change).