Update minimum Error Prone version and Guava version (#843)
Fixes #842. Updating the minimum supported Error Prone version to 2.10.0
allows for some code cleanup in tests.

We also update our `README.md` (finally!) to mention the new minimum
version, to use the Gradle Error Prone plugin correctly, to mention the
Gradle NullAway plugin, and to not mention so many specific versions so
it doesn't get out of date so quickly in the future. This fixes #466.