JSpecify: initial handling of generic enclosing types for inner classes (#837)
This is a step towards fixing #836. We now properly handle the case of a
generic enclosing type for an inner class type, both in checking
matching of type argument nullability and in pretty-printing for error
messages. We still don't handle `NewClassTree` types correctly which is
why #836 is not yet fixed.

Also fixes a bug in where generics checks were performed for `return`
statements. Previously we would only check generic type arguments if the
top-level nullability of the return type was `@NonNull`.