Model Lombok-generated equals methods as having a @Nullable parameter (#874)
See the new test in `UsesDTO`. Without this change, NullAway reports an
incorrect warning for this code that passes a `@Nullable` parameter to a
Lombok-generated `equals()` method.

The real logic change in this PR is very small. Most of the changes are
required to make a `VisitorState` object available in a `Handler`
method.