Extend library models to mark fields as nullable (#878)
This PR extends `LibraryModels` to add support for marking fields
`@Nullable`. This feature is required to enable
[Annotator](https://github.com/ucr-riple/NullAwayAnnotator) index impact
of making fields `@Nullable` on downstream dependencies.

Please note, since this feature is mostly going to be used while
indexing impacts on downstream dependencies, this PR does not focus on
the optimality of the implementation. Also the working set in iterations
is not expected to be large. We can optimize the implementation in a
follow up PR if needed.

---------

Co-authored-by: Manu Sridharan <msridhar@gmail.com>