Fix bug with implicit equals() methods in interfaces (#898)
Fixes #897 

This fixes a regression in our handling of implicit `equals()` methods
in interfaces when building on JDK 21. I see this as an interim fix,
until we can fix NullAway to properly always assume / enforce that the
parameter of `equals()` is `@Nullable`. But, I think we should fix the
regression in the short term.