Update instructions for Android and our sample app (#900)
Fixes #891. It's hard to add a regression test for this, but I manually
tested that with the new configuration, NullAway errors are detected in
the sample app. After this lands, I'd like to update the readme further
to refer to this particular commit, so users can see the changes we
needed to make.