JSpecify: read upper bound annotations from bytecode and add tests (#1004)
We add support for reading upper bound annotations from bytecodes for
JSpecify mode generics checks. Many more types of annotations need to be
read with special handling on JDK versions earlier than 22, due to
https://bugs.openjdk.org/browse/JDK-8225377. For now, we disable some
tests on earlier JDKs. We may add this special handling in the future,
but it will be a significant amount of work. We are still hoping the
relevant javac fix gets backported, which will fix this for us; see
https://github.com/jspecify/jspecify/issues/365.