Create basic unit tests for library model generation (#1031)
This is in preparation for #1006. When working on that PR I realized we
really need some unit tests to more easily debug various scenarios
(right now all we have are integration tests). Plus this will make our
code coverage reports more useful. This adds some infrastructure for
unit tests and a couple of basic tests; we will add more as we go. Also
fix a couple minor issues in `LibraryModelGenerator`.