Partial handling for restrictive annotations on varargs in unannotated code (#1029)
Partially addresses #1027. Most importantly, handles the case with
unannotated methods and _no_ restrictive annotations; we were getting
warnings in some cases with the main branch, which impacts backward
compatibility.

I started implementing full support for restrictive annotations on
varargs parameters and then realized it would take some refactoring
(since there can be a restrictive annotation on the varargs array itself
or on its contents, and we have no way to express this in the current
code structure). This PR contains some incomplete code towards that
change, which I figured does no harm so I left it in for now. If desired
I can remove it and put it in a separate PR.